# docs/swagger.yaml
basePath: /api/
definitions:
...
  db.Estudiante:
    properties:
      apellido:
        type: string
      correo:
        type: string
      estudianteID:
        type: integer
      fechaRegistro:
        $ref: '#/definitions/pgtype.Timestamp'
      nombre:
        type: string
      programaAcademico:
        type: string
      semestre:
        $ref: '#/definitions/pgtype.Int4'
      ti:
        $ref: '#/definitions/pgtype.Int4'
    type: object
...
  handler.CreateEstudianteRequest:
    properties:
      apellido:
        example: Perez
        type: string
      correo:
        example: juan.perez@urosario.edu.co
        type: string
      nombre:
        example: Juan
        type: string
      programa_academico:
        example: Ingeniería de Sistemas
        type: string
      semestre:
        example: 5
        type: integer
      ti:
        example: 1000123456
        type: integer
    type: object
  handler.CreateEstudianteResponse:
    properties:
      estudiante_id:
        type: integer
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: Descriptive error message
        type: string
    type: object
...
  handler.UpdateEstudianteRequest:
    properties:
      apellido:
        example: Perez
        type: string
      correo:
        example: juan.perez@urosario.edu.co
        type: string
      nombre:
        example: Juan Carlos
        type: string
      programa_academico:
        example: Ingeniería de Sistemas
        type: string
      semestre:
        example: 6
        type: integer
      ti:
        example: 1000123456
        type: integer
    type: object
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Int4:
    properties:
      int32:
        type: integer
      valid:
        type: boolean
    type: object
...
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
host: matwa.tail013c29.ts.net
info:
  contact: {}
  description: This is the API for my project.
  termsOfService: http://swagger.io/terms/
  title: Proyecto de Ingenieria de  Datos API
  version: "1.0"
paths:
...
  /v1/estudiantes:
    get:
      description: Retrieves students filtered by academic program.
      parameters:
      - description: Academic program
        in: query
        name: programa
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved estudiantes
          schema:
            items:
              $ref: '#/definitions/db.Estudiante'
            type: array
        "500":
          description: Failed to retrieve estudiantes
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List Estudiantes by Program
      tags:
      - Estudiantes
    post:
      consumes:
      - application/json
      description: Creates a new student record.
      parameters:
      - description: Estudiante Data
        in: body
        name: estudiante
        required: true
        schema:
          $ref: '#/definitions/handler.CreateEstudianteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created student
          schema:
            $ref: '#/definitions/handler.CreateEstudianteResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to create student
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create Estudiante
      tags:
      - Estudiantes
  /v1/estudiantes/{id}:
    delete:
      description: Deletes a student by their ID.
      parameters:
      - description: Estudiante ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted estudiante
        "400":
          description: Invalid estudiante ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to delete estudiante
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Estudiante
      tags:
      - Estudiantes
    get:
      description: Retrieves a specific student by their ID.
      parameters:
      - description: Estudiante ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved estudiante
          schema:
            $ref: '#/definitions/db.Estudiante'
        "400":
          description: Invalid estudiante ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Estudiante not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to retrieve estudiante
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get Estudiante by ID
      tags:
      - Estudiantes
    put:
      consumes:
      - application/json
      description: Updates an existing student.
      parameters:
      - description: Estudiante ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Estudiante Data
        in: body
        name: estudiante
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateEstudianteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated estudiante
          schema:
            $ref: '#/definitions/db.Estudiante'
        "400":
          description: Invalid request body or estudiante ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Estudiante not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to update estudiante
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update Estudiante
      tags:
      - Estudiantes
...
schemes:
- https
swagger: "2.0"